Move Radio's styled components into `.styled` file
