+++ frontend/src/metabase/components/Radio.jsx
@@ -1,12 +1,16 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import _ from "underscore";
-import cx from "classnames";
-import styled from "styled-components";
-import { space } from "styled-system";
 
-import { color, lighten } from "metabase/lib/colors";
 import Icon from "metabase/components/Icon";
+import {
+  BubbleList,
+  BubbleItem,
+  NormalList,
+  NormalItem,
+  UnderlinedList,
+  UnderlinedItem,
+} from "./Radio.styled";
 
 const propTypes = {
   name: PropTypes.string,
@@ -115,64 +119,3 @@ Radio.propTypes = propTypes;
 Radio.defaultProps = defaultProps;
 
 export default Radio;
-
-// BASE components all variants inherit from
-const BaseList = styled.ul`
-  display: flex;
-  flex-direction: ${props => (props.vertical ? "column" : "row")};
-`;
-const BaseItem = styled.li.attrs({
-  mr: props => (!props.vertical && !props.last ? props.xspace : null),
-  mb: props => (props.vertical && !props.last ? props.yspace : null),
-  "aria-selected": props => props.selected,
-})`
-  ${space}
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  :hover {
-    color: ${props =>
-      !props.showButtons && !props.selected ? color("brand") : null};
-  }
-`;
-BaseItem.defaultProps = {
-  xspace: 3,
-  yspace: 1,
-};
-
-// NORMAL
-const NormalList = styled(BaseList).attrs({
-  className: props => cx(props.className, { "text-bold": !props.showButtons }),
-})``;
-const NormalItem = styled(BaseItem)`
-  color: ${props => (props.selected ? color("brand") : null)};
-`;
-
-// UNDERLINE
-const UnderlinedList = styled(NormalList)``;
-const UnderlinedItem = styled(NormalItem)`
-  border-bottom: 3px solid transparent;
-  border-color: ${props => (props.selected ? color("brand") : null)};
-`;
-UnderlinedItem.defaultProps = {
-  py: 2,
-};
-
-// BUBBLE
-const BubbleList = styled(BaseList)``;
-const BubbleItem = styled(BaseItem)`
-  font-weight: 700;
-  border-radius: 99px;
-  color: ${props => (props.selected ? color("white") : color("brand"))};
-  background-color: ${props =>
-    props.selected ? color("brand") : lighten("brand")};
-  :hover {
-    background-color: ${props => !props.selected && lighten("brand", 0.38)};
-    transition: background 300ms linear;
-  }
-`;
-BubbleItem.defaultProps = {
-  xspace: 1,
-  py: 1,
-  px: 2,
-};
